Reconstruct system of constants
