+++ Car/src/activity/journey/journey-activity/journey-applicant/JourneyApplicant.tsx
@@ -5,7 +5,7 @@ import User from "../../../../../models/user/User";
 import AvatarLogo from "../../../../components/avatar-logo/AvatarLogo";
 import JourneyApplicantStyle from "./JourneyApplicantStyle";
 import Indicator from "../../../../components/activity-indicator/Indicator";
-import { SINGLE_ELEMENT_COLLECTION_LENGTH } from "../../../../constants/Constants";
+import { SINGLE_ELEMENT_COLLECTION_LENGTH } from "../../../../constants/GeneralConstants";
 import DM from "../../../../components/styles/DM";
 
 const JourneyApplicant = (props: {route: {params: { userId: number }}}) => {
