+++ Car/src/activity/my-profile/my-profile-activity/cars/car-activity/edit-cars/EditCars.tsx
@@ -22,7 +22,8 @@ import CarDropDownPicker from "../../../../../../components/car-drop-down-picker
 import CarTextInput from "../../../../../../components/car-text-input/CarTextInput";
 import EditCarsStyle from "./EditCarsStyle";
 import DM from "../../../../../../components/styles/DM";
-import { FIRST_ELEMENT_INDEX, MAX_PLATE_NUMBER_LENGTH, MIN_PLATE_NUMBER_LENGTH } from "../../../../../../constants/Constants";
+import { MAX_PLATE_NUMBER_LENGTH, MIN_PLATE_NUMBER_LENGTH } from "../../../../../../constants/CarConstants";
+import { FIRST_ELEMENT_INDEX } from "../../../../../../constants/GeneralConstants";
 import AuthContext from "../../../../../../components/auth/AuthContext";
 import CreateCarViewModel from "../../../../../../../models/car/CreateCarViewModel";
