+++ Car/src/components/avatar-logo-title/AvatarLogoTitle.tsx
@@ -3,7 +3,8 @@ import { Text, View } from "react-native";
 import AvatarLogo from "../avatar-logo/AvatarLogo";
 import AuthContext from "../auth/AuthContext";
 import AvatarLogoTitleStyle from "./AvatarLogoTitleStyle";
-import { SINGLE_ELEMENT_COLLECTION_LENGTH, USER_STATE_CHANGE_EVENT_NAME } from "../../constants/Constants";
+import { USER_STATE_CHANGE_EVENT_NAME } from "../../constants/ProfileConstants";
+import { SINGLE_ELEMENT_COLLECTION_LENGTH } from "../../constants/GeneralConstants";
 import User from "../../../models/user/User";
 import DM from "../styles/DM";
 import { EventRegister } from "react-native-event-listeners";
