+++ Car/src/components/journey-card/JourneyCard.tsx
@@ -5,7 +5,7 @@ import * as navigation from "../navigation/Navigation";
 import AvatarLogo from "../avatar-logo/AvatarLogo";
 import moment from "moment";
 import AuthContext from "../auth/AuthContext";
-import { FIRST_ELEMENT_INDEX, LAST_INDEX_CORRECTION } from "../../constants/Constants";
+import { FIRST_ELEMENT_INDEX, LAST_INDEX_CORRECTION } from "../../constants/GeneralConstants";
 import DM from "../styles/DM";
 import Journey from "../../../models/journey/Journey";
