+++ Car/src/constants/DimensionConstants.tsx
@@ -1,0 +1,2 @@
+export const MILLISECONDS_IN_MONTH = 2629800000;
+export const MILLISECONDS_IN_MINUTES = 60000;
