+++ Car/src/constants/GeneralConstants.tsx
@@ -1,0 +1,12 @@
+export const NOT_EXISTING_ELEMENT_INDEX = -1;
+export const FIRST_ELEMENT_INDEX = 0;
+export const SECOND_ELEMENT_INDEX = 1;
+export const THIRD_ELEMENT_INDEX = 2;
+export const FOURTH_ELEMENT_INDEX = 3;
+export const SECOND_FROM_END_ELEMENT_INDEX = -2;
+export const THIRD_FROM_END_ELEMENT_INDEX = -3;
+export const LAST_INDEX_CORRECTION = 1;
+export const EMPTY_COLLECTION_LENGTH = 0;
+export const SINGLE_ELEMENT_COLLECTION_LENGTH = 1;
+export const THREE_ELEMENT_COLLECTION_LENGTH = 3;
+export const DELETE_COUNT = 1;
