+++ Car/src/constants/JourneyConstants.tsx
@@ -1,0 +1,20 @@
+import { getStatusBarHeight } from "react-native-status-bar-height";
+
+export const JOURNEY_MORE_OPTIONS_POPUP_HEIGHT = 280;
+export const CREATE_JOURNEY_MORE_OPTIONS_POPUP_HEIGHT = 200;
+export const REQUEST_RIDE_POPUP_HEIGHT = 195;
+export const JOURNEY_CONTENT_HEIGHT = 201 + getStatusBarHeight();
+export const MIN_JOURNEY_PAGE_POPUP_HEIGHT = 0;
+export const MEDIUM_JOURNEY_PAGE_POPUP_HEIGHT = 262 + getStatusBarHeight();
+export const MAX_JOURNEY_PAGE_POPUP_HEIGHT = 663 + getStatusBarHeight();
+export const MIN_JOURNEY_REQUEST_PAGE_POPUP_HEIGHT = 0;
+export const MEDIUM_JOURNEY_REQUEST_PAGE_POPUP_HEIGHT = 147;
+export const MAX_JOURNEY_REQUEST_PAGE_POPUP_HEIGHT = 412;
+export const INITIAL_PASSENGERS_COUNT = 1;
+export const NUMBER_OF_STOPS_LIMIT = 7;
+export const INITIAL_TIME = 0;
+export const INITIAL_ROUTE_DISTANCE = 0;
+export const DEFAULT_AVAILABLE_SEATS_COUNT = 4;
+export const LEFT_PADDING_FOR_FROM_PLACEHOLDER = 67;
+export const LEFT_PADDING_FOR_TO_PLACEHOLDER = 45;
+export const LEFT_PADDING_FOR_VIA_PLACEHOLDER = 50;
