+++ Car/src/constants/MessageConstants.tsx
@@ -1,0 +1,5 @@
+export const MESSAGE_SEARCH_START_AFTER_SYMBOLS_NUMBER = 2;
+export const MESSAGE_SEARCH_INPUT_SYMBOL_LIMIT = 25;
+export const CHAT_POPUP_HEIGHT = 135;
+export const FIRST_LOADING_MESSAGES = 0;
+export const COUNT_OF_MESSAGES_TO_LOAD = 0;
