+++ Car/src/constants/ProfileConstants.tsx
@@ -1,0 +1,2 @@
+export const MAX_PHOTO_FILE_SIZE = 7e+6;
+export const USER_STATE_CHANGE_EVENT_NAME = "onUserStateChange";
