+++ Car/src/constants/StylesConstants.tsx
@@ -1,0 +1,17 @@
+import { Dimensions } from "react-native";
+
+export const DROP_DOWN_MAX_HEIGHT = Dimensions.get("window").height / 4;
+export const GRADIENT_START = 0;
+export const GRADIENT_END = 1;
+export const ZERO_OPACITY = 0;
+export const ZERO_MARGIN = 0;
+export const ZERO_COORDINATE = 0;
+export const HALF_OPACITY = 0.5;
+export const MAX_OPACITY = 1;
+export const MAX_POPUP_POSITION = 0;
+export const MIN_POPUP_POSITION = 1;
+export const MIN_POPUP_HEIGHT = 0;
+export const POPUP_HEIGHT_WITH_USER_IMAGE = 188;
+export const POPUP_HEIGHT_WITHOUT_USER_IMAGE = 143;
+export const AVATAR_LOGO_SIZE_TO_TEXT_RATIO = 2.5;
+export const AVATAR_LOGO_SIZE_TO_PADDING_RATIO = 14;
